Move examples to separate files - second part
